+++ main_test.go
@@ -5,9 +5,18 @@ import (
 	"io/ioutil"
 	"net/http"
 
+	"github.com/rightscale/rsc/cmd"
+	"gopkg.in/alecthomas/kingpin.v2"
+
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
+
+type timeoutError struct{ error }
+
+func (e *timeoutError) Error() string   { return "i/o timeout" }
+func (e *timeoutError) Timeout() bool   { return true }
+func (e *timeoutError) Temporary() bool { return true }
 
 var _ = Describe("Main", func() {
 
@@ -44,6 +53,36 @@ var _ = Describe("Main", func() {
 			})
 		})
 
+	})
+
+	Context("retry option", func() {
+		var app = kingpin.New("rsc", "rsc - tests")
+		var retries = 6
+		var cmdLine = cmd.CommandLine{Retry: retries}
+		It("retries if error on API call occurs", func() {
+			counter := 0
+			doAPIRequest = func(string, *cmd.CommandLine) (*http.Response, error) {
+				counter += 1
+				return nil, &timeoutError{}
+			}
+			ExecuteCommand(app, &cmdLine)
+			Ω(counter).Should(Equal(1 + retries))
+		})
+
+		It("doesn't retry more than necessary", func() {
+			counter := 0
+			doAPIRequest = func(string, *cmd.CommandLine) (*http.Response, error) {
+				counter += 1
+				if counter < 3 {
+					return &http.Response{StatusCode: 503}, nil
+				} else {
+					return nil, nil
+				}
+			}
+			_, err := ExecuteCommand(app, &cmdLine)
+			Ω(err).ShouldNot(HaveOccurred())
+			Ω(counter).Should(BeNumerically("<", 1+retries))
+		})
 	})
 
 })